Cleaning up some build warnings.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/trunk@1594 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Analysis/analysis/trunk/sourcefitting/RadioSource.h
@@ -244,7 +244,7 @@ namespace askap
 
 	float itsChisq;
 	float itsRMS;
-	float itsNDoF;
+	int itsNDoF;
 
 	/// @brief The min & max points of the box, taking into account the borders of the data array
 	std::vector<std::pair<long,long> > itsBoxMargins;
